requirements.txt
discord.py
edge-tts
PyNaCl
